backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Configuration
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0

# Cloud Storage
cloudinary==1.41.0

# HTTP requests
requests==2.32.3
httpx==0.27.2

# Monitoring (optional)
sentry-sdk==2.14.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
